Allow VLIME to continue upon reading NUL characters.
Handling NULs in JSON is a bit cumbersome in nvim; while possible,
it would need some special code. So as a first fix we just substitute
NULs in the data stream, we shouldn't see such in the REPL anyway.

As for why we need to handle them at all, just eval
    (make-string 4)