Change `CurTopExprPos` to use `SearchParenPos`
Given the following buffer:

    (defun hello-world()
      (format t "Hello, world!"))

    (defun open-parenthesis-p (ch)
      "Some documentation containing unbalanced ("
      (char= ch #\())

    (defun close-parenthesis-p (ch)
      "Some documentation containing unbalanced )"
      (char= ch #\)))

Placing the cursor onto the `#\(` character literal, and running `echo
vlime#ui#CurTopExprPos('begin')` correctly logs: `[4, 1]`; however, if
we pass in `'end'` we get `[10, 17]` instead of the expected `[6, 17]`.
Similarly, if we place the cursor onto the `#\)` character literal and
run `echo vlime#ui#CurTopExprPos('end')` we get: `[10, 17]`, but if we
use `'begin'` we get `[4, 1]` instead of `[8, 1]`.  Switching to our
`searchpairpos()` wrapper seems to be fixing the problem.