Change `CurOperator` to use `SearchParenPos`
Given the following buffer:

    (defun hello-world()
      (format t "Hello, world!"))

    (defun open-parenthesis-p (ch)
      "Some documentation containing unbalanced ("
      (char= ch #\())

    (defun close-parenthesis-p (ch)
      "Some documentation containing unbalanced )"
      (char= ch #\)))

Placing the cursor onto the `#\(` character literal, and running `echo
vlime#ui#CurOperator()` returns an empty string, and interestingly
enough adding `skip` to `searchpairpos()` does not seem to make much
difference.  Why? Because the function calls `vlime#ui#CurExpr()` which
returns `()` instead of the expected `(char= ch #\())`.  Why is
`vlime#ui#CurExpr()` returning `()`? Because we are re-using the same
logic of matchparen.vim, which intentionally does _not_ try to skip over
certain syntax regions if the position the was in at the time the search
operation was issued happened to be within one of those syntax regions
already (i.e. it will highlight parentheses inside comments).

In any case, since we have to deal with incomplete expressions anyway,
I am proposing that we stop calling `vlime#ui#CurExpr()` here, and
simply use the `searchpairpos()` wrapper instead.