Change `SurroundingOperator` to use `SearchParenPos`
Given the following buffer:

    (defun hello-world()
      (format t "Hello, world!"))

    (defun open-parenthesis-p (ch)
      "Some documentation containing unbalanced ("
      (char= ch #\())

    (defun close-parenthesis-p (ch)
      "Some documentation containing unbalanced )"
      (char= ch #\)))

Placing the cursor onto the closing parenthesis right after the `#\(`
character literal and calling this function outputs an empty string
instead of `char=`.  The same happens if we place the cursor at the
closing parenthesis right after the `#\)` character literal of the
second function definition.  Using our `searchpairpos()` wrapper here,
seems to be fixing the problem.