Change `CurArgPos` to use `SearchParenPos`
Given the following buffer:

    (defun open-parenthesis-yoda-p (ch)
      "Some documentation containing unbalanced ("
      (char= #\( ch))

    (defun close-parenthesis-yoda-p (ch)
      "Some documentation containing unbalanced )"
      (char= #\) ch))

Placing the cursor onto the `ch` symbol right after the `#\(` character
literal and calling `vlime#ui#CurArgPos()` outputs `0` instead of the
`2`.  Similarly, doing the same from the `ch` symbol right after the
`#\)` character literal outputs `5` instead of `2`.  To fix this, not
only we had to use our `searchpairpos()` wrapper, but also change the
_parser_ to try to deal with character literals.