Access swank symbols dynamically via SYMBOL-VALUE
This makes it possible to LOAD load-vlime.lisp without swank being yet
loaded into the Lisp image.  And with this, it should now be possible to
_cleanly_ LOAD Vlime/Swank with the following 3 steps:

1. You LOAD load-vlime.lisp
2. You call SWANK-LOADER:INIT, e.g., `(dyn-call "SWANK-LOADER" "INIT" :load-contribs t)`
3. You call VLIME:MAIN

Previously, one had to call SWANK-LOADER:INIT first, then LOAD
load-vlime.lisp, and then call VLIME:MAIN; same steps, but in different
order.  The problem with that is that load-vlime.lisp would internally
reload the :VLIME system, which in turn would trigger a reload of the
:SWANK, which would result in the compiler emitting a ton of warnings
about symbol redefinitions happening inside SWANK* packages.